pyproject.toml
[tool.poetry]
name = "model-service"
version = "0.1.0"
description = ""
authors = ["Ivan Krasnoshchok <ivankras67@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pika = "^1.3.1"
mongoengine = "^0.27.0"
attrs = "^22.2.0"
pydantic = "^1.10.7"
pandas = "^2.0.0"
tensorflow-macos = "^2.12.0"
boto3 = "^1.26.113"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
